Add functional configuration option to `NewStubRequest`.
Previously we provided this functionality by java style function
chaining, but that always felt a little unidiomatic. This replaces with
the functional configuration style I've recently been using everywhere
that is described here: https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis

The previous code is maintained as these changes are compatible; I've
just marked the chaining functions as deprecated.